Fixed issue with null values on XY charts due to toArray() fixes.

+++ src/org/achartengine/compat/PathMeasure.java
@@ -20,7 +20,8 @@ public class PathMeasure {
         forceClosed = b;
     }
     
-    public int getLength(){
+    public float getLength(){
+        
         return path.path().length();
     }
